move handler.go outside client/
